SMBG_carte_starter/requirements.txt
streamlit>=1.36
pandas>=2.2
openpyxl>=3.1
requests>=2.31
pyyaml>=6.0
phonenumberslite>=8.13
folium>=0.16
streamlit-folium>=0.20
reportlab>=3.6
supabase>=2.5
python-dotenv>=1.0
Pillow>=10.0
